// app-master-detail/src/app/interfaces/customer.interface.ts
export interface Customer {
  customerId?: number;
  firstName?: string;
  pSurname?: string;
  mSurname?: string;
  email?: string;
  company?: string;
  phone?: string;
  address?: string;
  type?: string;
  id?: number;//si
  nombre?: string;//si
  apellidos?: string;//si
  celular?: number;//si

}